requirements.txt
requests>=2.7.0
six
humanize
fire
docopt
progress>=1.3
retry>=0.9.2
whichcraft
pure-python-adb
logzero
# remove pillow because pillow can't installed on QPython
# pillow
pillow ; sys_platform == "win32" or sys_platform == "darwin"
lxml ; sys_platform == "win32" or sys_platform == "darwin"
